use id count instead of uuid in gdpr test to avoid random test failures
