+++ frontend/kesaseteli/employer/browser-tests/actions/application.actions.ts
@@ -1,5 +1,5 @@
 import type { SuomiFiData } from '@frontend/shared/browser-tests/actions/login-action';
-import { fakeApplication } from '@frontend/shared/src/__tests__/utils/fake-objects';
+import FakeObjectFactory from '@frontend/shared/src/__tests__/utils/FakeObjectFactory';
 import isRealIntegrationsEnabled from '@frontend/shared/src/flags/is-real-integrations-enabled';
 import type Application from '@frontend/shared/src/types/application';
 import Employment from '@frontend/shared/src/types/employment';
@@ -15,6 +15,8 @@ import { doEmployerLogin } from './employer-header.actions';
 
 type UserAndApplicationData = Application & SuomiFiData;
 
+const fakeObjectFactory = new FakeObjectFactory();
+
 export const fillStep1Form = async (
   t: TestController,
   application: Application
@@ -26,22 +28,12 @@ export const fillStep1Form = async (
     street_address,
     bank_account_number,
     contact_person_phone_number,
-    is_separate_invoicer,
-    invoicer_name,
-    invoicer_email,
-    invoicer_phone_number,
   } = application;
   await step1Form.actions.fillContactPersonName(contact_person_name);
   await step1Form.actions.fillContactPersonEmail(contact_person_email);
   await step1Form.actions.fillContactPersonPhone(contact_person_phone_number);
   await step1Form.actions.fillStreetAddress(street_address);
   await step1Form.actions.fillBankAccountNumber(bank_account_number);
-  if (is_separate_invoicer) {
-    await step1Form.actions.selectSeparateInvoicerCheckbox();
-    await step1Form.actions.fillInvoicerName(invoicer_name);
-    await step1Form.actions.fillInvoicerEmail(invoicer_email);
-    await step1Form.actions.fillInvoicerPhone(invoicer_phone_number);
-  }
 };
 
 export const removeStep2ExistingForms = async (
@@ -116,13 +108,8 @@ export const loginAndfillApplication = async (
   const urlUtils = getUrlUtils(t);
   const suomiFiData = await doEmployerLogin(t);
   const wizard = await getWizardComponents(t);
-  const applicationId =
-    await urlUtils.expectations.urlChangedToApplicationPage();
-  const application = fakeApplication(
-    applicationId,
-    suomiFiData?.company,
-    true
-  );
+  await urlUtils.expectations.urlChangedToApplicationPage();
+  const application = fakeObjectFactory.fakeApplication(suomiFiData?.company);
   // if there is existing draft application on step 2 or 3, then move to step 1.
   await wizard.actions.clickGoToStep1Button();
   if (toStep >= 1) {
