+++ frontend/kesaseteli/employer/browser-tests/application-page/summary.components.ts
@@ -8,7 +8,6 @@ import Application from '@frontend/shared/src/types/application';
 import Company from '@frontend/shared/src/types/company';
 import ContactInfo from '@frontend/shared/src/types/contact-info';
 import Employment from '@frontend/shared/src/types/employment';
-import Invoicer from '@frontend/shared/src/types/invoicer';
 import { convertToUIDateFormat } from '@frontend/shared/src/utils/date.utils';
 import { friendlyFormatIBAN } from 'ibantools';
 import TestController from 'testcafe';
@@ -51,7 +50,7 @@ export const getSummaryComponents = async (t: TestController) => {
       companyHeading() {
         return findEmployerField('company-heading');
       },
-      employerField(field: keyof Company | keyof ContactInfo | keyof Invoicer) {
+      employerField(field: keyof Company | keyof ContactInfo) {
         return findEmployerField(`${String(field)}`);
       },
     };
@@ -80,7 +79,7 @@ export const getSummaryComponents = async (t: TestController) => {
       },
       async isFulFilledWith(application: Application) {
         const expectFieldHasValue = async (
-          field: keyof ContactInfo | keyof Invoicer,
+          field: keyof ContactInfo,
           value?: string | number
         ) =>
           expectElementHasValue(
@@ -96,11 +95,6 @@ export const getSummaryComponents = async (t: TestController) => {
           'bank_account_number',
           friendlyFormatIBAN(application.bank_account_number) ?? ''
         );
-        if (application.is_separate_invoicer) {
-          await expectFieldHasValue('invoicer_name');
-          await expectFieldHasValue('invoicer_email');
-          await expectFieldHasValue('invoicer_phone_number');
-        }
       },
     };
     await expectations.isPresent();
