+++ frontend/kesaseteli/employer/public/locales/en/common.json
@@ -97,10 +97,6 @@
         "street_address": "Company street address",
         "bank_account_number": "ENG Tilinumero",
         "contact_person_phone_number": "Telephone number",
-        "is_separate_invoicer": "The billing is handled by an other person",
-        "invoicer_name": "Name of the billing contact person",
-        "invoicer_email": "Email of the billing contact person",
-        "invoicer_phone_number": "Phone number of the billing contact person",
         "employee_name": "Name of the employee",
         "employee_ssn": "Social security number",
         "summer_voucher_exception_reason": "Grade or birth year",
