+++ frontend/kesaseteli/employer/public/locales/fi/common.json
@@ -97,10 +97,6 @@
         "street_address": "Työpaikan lähiosoite",
         "bank_account_number": "Tilinumero",
         "contact_person_phone_number": "Yhteyshenkilön puhelinnumero",
-        "is_separate_invoicer": "Laskutusta hoitaa yrityksessä eri yhteyshenkilö",
-        "invoicer_name": "Laskutuksen yhteyshenkilön nimi",
-        "invoicer_email": "Laskutuksen yhteyshenkilön sähköposti",
-        "invoicer_phone_number": "Laskutuksen yhteyshenkilön puhelinnumero",
         "employee_name": "Työntekijän nimi",
         "employee_ssn": "Henkilötunnus",
         "summer_voucher_exception_reason": "Vuosiluokka tai syntymävuosi",
