+++ frontend/kesaseteli/employer/public/locales/sv/common.json
@@ -97,10 +97,6 @@
         "street_address": "Arbetsplatsens gatuadress",
         "bank_account_number": "SWE Tilinumero",
         "contact_person_phone_number": "Telefonnummer",
-        "is_separate_invoicer": "Faktureringen sköts av en annan person",
-        "invoicer_name": "Kontakt person för faktureringen",
-        "invoicer_email": "E-posten för kontaktperson för fakturan",
-        "invoicer_phone_number": "Telefonnummer för kontaktperson för fakturan ",
         "employee_name": "Arbetstagarens namn",
         "employee_ssn": "Personnummer",
         "summer_voucher_exception_reason": "Årsklass eller födelseår",
