+++ frontend/kesaseteli/employer/src/components/header/__tests__/Header.test.tsx
@@ -7,9 +7,11 @@ import headerApi from 'kesaseteli-shared/__tests__/utils/component-apis/header-a
 import renderComponent from 'kesaseteli-shared/__tests__/utils/components/render-component';
 import { getBackendUrl } from 'kesaseteli-shared/backend-api/backend-api';
 import React from 'react';
-import { fakeUser } from 'shared/__tests__/utils/fake-objects';
+import FakeObjectFactory from 'shared/__tests__/utils/FakeObjectFactory';
 import { waitFor } from 'shared/__tests__/utils/test-utils';
 import { DEFAULT_LANGUAGE, SUPPORTED_LANGUAGES } from 'shared/i18n/i18n';
+
+const fakeObjectFactory = new FakeObjectFactory();
 
 describe('frontend/kesaseteli/employer/src/components/header/Header.tsx', () => {
   it('Redirects to backend when clicked login button', async () => {
@@ -26,7 +28,7 @@ describe('frontend/kesaseteli/employer/src/components/header/Header.tsx', () => 
   });
 
   it('Userdata is cleared when clicked logout button', async () => {
-    const user = fakeUser();
+    const user = fakeObjectFactory.fakeUser();
     expectAuthorizedReply(user);
     const spyRouterPush = jest.fn();
     const { queryClient } = renderComponent(<Header />, {
