--- frontend/shared/src/types/invoicer.ts
@@ -1,7 +1,0 @@
-type Invoicer = {
-  invoicer_email?: string;
-  invoicer_name?: string;
-  invoicer_phone_number?: string;
-};
-
-export default Invoicer;
