+++ frontend/shared/src/utils/mask-gdpr-data.ts
@@ -58,9 +58,6 @@ export const ATTRIBUTES_TO_MASK = [
   'contact_person_phone_number',
   'street_address',
   'bank_account_number',
-  'invoicer_email',
-  'invoicer_name',
-  'invoicer_phone_number',
   'last_name',
   'social_security_number',
   'email',
